backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
websockets==12.0
redis==5.0.1
celery==5.3.4
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
httpx==0.25.2
